lots of changes, lost track of what they were but it works

+++ internals/common.py
@@ -22,19 +22,19 @@ def get_ground_truth(truth_df: pd.DataFrame, epoch_df: pd.DataFrame) -> tuple[to
         gt_row['LatitudeDegrees'], gt_row['LongitudeDegrees']
     )
     # Convert to tensors
-    pos_truth = torch.tensor(pos_truth, dtype=torch.float32, device=get_device())
-    vel_truth = torch.tensor(vel_truth, dtype=torch.float32, device=get_device())
+    pos_truth = torch.tensor(pos_truth, dtype=torch.float64, device=get_device())
+    vel_truth = torch.tensor(vel_truth, dtype=torch.float64, device=get_device())
     return pos_truth, vel_truth
 
 def compute_pos_torch(epoch_df: pd.DataFrame, pr_weights: torch.Tensor | None, pr_correction: torch.Tensor | None, prr_weights: torch.Tensor | None, curr_pos: dict | None) -> dict:
     # Grab the required columns and convert to tensors
-    pr = torch.tensor(epoch_df[PR_COL].to_numpy(), dtype=torch.float32, device=get_device())
+    pr = torch.tensor(epoch_df[PR_COL].to_numpy(), dtype=torch.float64, device=get_device())
     if pr_correction is not None:
         pr = pr + pr_correction
         
-    prr = torch.tensor(epoch_df[PRR_COL].to_numpy(), dtype=torch.float32, device=get_device())
-    sat_pos = torch.tensor(epoch_df[SAT_POS_COLS].to_numpy(), dtype=torch.float32, device=get_device())
-    sat_vel = torch.tensor(epoch_df[SAT_VEL_COLS].to_numpy(), dtype=torch.float32, device=get_device())
+    prr = torch.tensor(epoch_df[PRR_COL].to_numpy(), dtype=torch.float64, device=get_device())
+    sat_pos = torch.tensor(epoch_df[SAT_POS_COLS].to_numpy(), dtype=torch.float64, device=get_device())
+    sat_vel = torch.tensor(epoch_df[SAT_VEL_COLS].to_numpy(), dtype=torch.float64, device=get_device())
     # Compute updated position estimate
     curr_pos = gp.position_torch(pr, prr, sat_pos, sat_vel, Wx=pr_weights, Wv=prr_weights, prev_estimate=curr_pos)
     return curr_pos
@@ -58,12 +58,12 @@ def compute_residual_matrix(epoch_df: pd.DataFrame) -> tuple[np.ndarray, np.ndar
 
     pr_residual_matrix = np.zeros(
         (n_sats, n_sats),
-        dtype=np.float32,
+        dtype=np.float64,
     )
 
     prr_residual_matrix = np.zeros(
         (n_sats, n_sats),
-        dtype=np.float32,
+        dtype=np.float64,
     )
 
     # Ensure unique satellite + signal identity
@@ -104,8 +104,8 @@ def compute_residual_matrix(epoch_df: pd.DataFrame) -> tuple[np.ndarray, np.ndar
         col_indices = [
             i for i, key in enumerate(sat_keys) if key != excluded_key
         ]
-        pr_residual_matrix[sat_num, col_indices] = pr_res.astype(np.float32)
-        prr_residual_matrix[sat_num, col_indices] = prr_res.astype(np.float32)
+        pr_residual_matrix[sat_num, col_indices] = pr_res.astype(np.float64)
+        prr_residual_matrix[sat_num, col_indices] = prr_res.astype(np.float64)
 
         # Diagonal explicitly set
         pr_residual_matrix[sat_num, sat_num] = 0.0
